% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_filesystem.R
\name{FilesystemStore}
\alias{FilesystemStore}
\title{Filesystem store}
\usage{
FilesystemStore(root)
}
\arguments{
\item{root}{Path to the root directory of the store. Will be created if it
does not exist.}
}
\value{
A \code{FilesystemStore} object.
}
\description{
A Zarr V3 store backed by a local filesystem directory. Keys map directly
to file paths relative to \code{root}. Intermediate directories are created
automatically on write.
}
