% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectstore-package.R
\docType{package}
\name{objectstore-package}
\alias{objectstore}
\alias{objectstore-package}
\title{objectstore: Object Store Interface}
\description{
A unified interface to object stores, inspired by the Zarr V3 specification. A set of S7 classes and methods to interact with object stores, such as the local filesystem or AWS S3.
}
\author{
\strong{Maintainer}: Hugo Gruson \email{hugo.gruson+R@normalesup.org}

Authors:
\itemize{
  \item Hugo Gruson \email{hugo.gruson+R@normalesup.org}
}

}
\keyword{internal}
