% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{Store}
\alias{Store}
\title{Abstract Zarr V3 store}
\usage{
Store()
}
\description{
Virtual base class for all store backends. Subclass this to implement a
concrete store (e.g. filesystem, in-memory, S3).
}
