% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{store_check_exist}
\alias{store_check_exist}
\title{Check if a key exists in the store}
\usage{
store_check_exist(store, ...)
}
\arguments{
\item{store}{A \code{ZarrStore} object.}

\item{key}{A string key.}
}
\value{
\code{TRUE} if \code{key} exists in \code{store}, \code{FALSE} otherwise.
}
\description{
Check if a key exists in the store
}
