% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{store_erase}
\alias{store_erase}
\title{Erase a single key/value pair from the store}
\usage{
store_erase(store, ...)
}
\arguments{
\item{store}{A \code{ZarrStore} object.}

\item{key}{A string key to erase.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Erase a single key/value pair from the store
}
\note{
Some stores (e.g. Zip archives) do not support deletion.
}
