% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{store_erase_prefix}
\alias{store_erase_prefix}
\title{Erase all keys with a given prefix}
\usage{
store_erase_prefix(store, ...)
}
\arguments{
\item{store}{A \code{ZarrStore} object.}

\item{prefix}{A string prefix. All keys beginning with this prefix are
erased.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Erase all keys with a given prefix
}
\note{
Some stores (e.g. Zip archives) do not support deletion.
}
