% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{store_erase_values}
\alias{store_erase_values}
\title{Erase a set of key/value pairs from the store}
\usage{
store_erase_values(store, ...)
}
\arguments{
\item{store}{A \code{ZarrStore} object.}

\item{keys}{A character vector of keys to erase.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Erase a set of key/value pairs from the store
}
\note{
Some stores (e.g. Zip archives) do not support deletion.
}
