% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{store_get}
\alias{store_get}
\title{Retrieve the value associated with a given key}
\usage{
store_get(store, ...)
}
\arguments{
\item{store}{A \code{ZarrStore} object.}

\item{key}{A string key.}
}
\value{
The value associated with \code{key}, or \code{NULL} if the key is absent.
}
\description{
Retrieve the value associated with a given key
}
