% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{store_list}
\alias{store_list}
\title{Retrieve all keys in the store}
\usage{
store_list(store, ...)
}
\arguments{
\item{store}{A \code{ZarrStore} object.}
}
\value{
A character vector of all keys.
}
\description{
Retrieve all keys in the store
}
