% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{store_list_prefix}
\alias{store_list_prefix}
\title{Retrieve all keys with a given prefix}
\usage{
store_list_prefix(store, ...)
}
\arguments{
\item{store}{A \code{ZarrStore} object.}

\item{prefix}{A string prefix ending with \code{/}. The behaviour is undefined if
\code{prefix} does not end with \code{/}.}
}
\value{
A character vector of keys that start with \code{prefix}. For example,
if the store contains \code{"a/b"}, \code{"a/c/d"}, and \code{"e/f/g"}, then
\code{store_list_prefix(store, "a/")} returns \code{c("a/b", "a/c/d")}.
}
\description{
Retrieve all keys with a given prefix
}
