% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{store_set}
\alias{store_set}
\title{Store a (key, value) pair}
\usage{
store_set(store, ...)
}
\arguments{
\item{store}{A \code{ZarrStore} object.}

\item{key}{A string key.}

\item{value}{The value to store.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Store a (key, value) pair
}
